Trenchtextures
TrenchTextures now change. Could use a much a simpler way of detecting
which on is needed.

the tries and catches need to be replaced by ifs. For some reason
they're causing HUGE fps drops. Like 60fps to 2 fps. in regards, the
try's and catches shouldnt be used like this anyway so needs to be
changed regardless.