squares are now striped red if blocked off
technically its working as intended, building is blocked if a square has
a counter value of 2000 and has no building at all place on it. this
includes concrete. If the player blocks of squares and builds concrete
on them its perfectly ok, if squares are blocked off and they're just
grass they need to correct it. From what i can see it is literally
impossible atm to make it so there's no path to the end.